+++ src/MotionBlockPyramid.h
@@ -196,7 +196,7 @@ private:
 
     template <int nLogPel, typename PixelType>
     void DoSearchMVs(const FramePyramidLevel &pSrcFrame, const FramePyramidLevel &pRefFrame,
-        SearchType st, int stp, int lambda, int lsad, int pnew,
+        SearchType st, int stp, int lambda, int64_t lsad, int pnew,
         int plevel, VECTOR *globalMVec, int fieldShift,
         int pzero, int pglobal, int64_t badSAD, int badrange, bool meander, bool tryMany, bool chroma) noexcept;
 
@@ -208,7 +208,7 @@ private:
     void EstimateGlobalMVDoubledFallback(VECTOR &globalMVec) const noexcept;
 public:
     void SearchMVs(const FramePyramidLevel &pSrcFrame, const FramePyramidLevel &pRefFrame,
-        SearchType st, int stp, int lambda, int lsad, int pnew,
+        SearchType st, int stp, int lambda, int64_t lsad, int pnew,
         int plevel, VECTOR *globalMVec, int fieldShift, bool useSatd,
         int pzero, int pglobal, int64_t badSAD, int badrange, bool meander, bool tryMany, bool chroma, int bytesPerSample);
 
@@ -286,7 +286,7 @@ public:
 
     void SearchMVs(const FramePyramid &pSrcGOF, const FramePyramid &pRefGOF,
         SearchType searchType, int nSearchParam, int nPelSearch, int nLambda,
-        int lsad, int pnew, int plevel, bool global, int fieldShift, bool useSatd,
+        int64_t lsad, int pnew, int plevel, bool global, int fieldShift, bool useSatd,
         int pzero, int pglobal, int64_t badSAD, int badrange, bool meander, bool tryMany,
         SearchType coarseSearchType, bool chroma);
